typo
git-svn-id: https://svn.r-project.org/R/trunk@35275 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/main/array.c
@@ -1087,7 +1087,7 @@ SEXP do_colsum(SEXP call, SEXP op, SEXP args, SEXP rho)
 	    int *Cnt = NULL, *c;
 	    rx = REAL(x);
 	    if (!keepNA && OP == 3) {
-		Cnt = (double *) alloca(n*sizeof(int));
+		Cnt = (int *) alloca(n*sizeof(int));
 		memset(Cnt, 0, n*sizeof(int));
 	    }
 	    memset(rans, 0, n*sizeof(double));
